Move tests to pipeline extra
